Fix 'ericomshield-setup.sh: line 354: docker: command not found'

+++ Setup/ericomshield-setup.sh
@@ -351,7 +351,7 @@ function install_docker() {
         apt-get -qq update
         echo "done"
         #Stop shield (if running)
-        if [ "$ES_RUN_DEPLOY" == true ] && [ $(docker stack ls | grep -c $STACK_NAME) -ge 1 ]; then
+        if [ "$ES_RUN_DEPLOY" == true ] && [ -x "/usr/bin/docker" ] && [ $(docker stack ls | grep -c $STACK_NAME) -ge 1 ]; then
            log_message "Stopping Ericom Shield for Update (Docker) (Downtime)"
            docker stack rm $STACK_NAME
         fi
